requirements.txt
pandas>=1.1.4
seaborn>=0.11.0
matplotlib>=3.2.2
numpy==1.19.2
opencv-python>=4.1.1
Pillow>=7.1.2
psutil
torch>=1.10.0
torchvision>=0.8.1
torchsummary
tqdm>=4.64.0
scikit-learn<=1.1.2

tensorboard>=2.4.1
